Move host expiration field to seperate component
