+++ src/pages/spawn/spawnHost/spawnHostModal/HostDetailsForm.tsx
@@ -4,12 +4,10 @@ import Button from "@leafygreen-ui/button";
 import Checkbox from "@leafygreen-ui/checkbox";
 import { Subtitle, Body } from "@leafygreen-ui/typography";
 import { Input, Select } from "antd";
-import { set } from "date-fns";
-import DatePicker from "components/DatePicker";
 import Icon from "components/icons/Icon";
 import { InputLabel } from "components/styles";
-import TimePicker from "components/TimePicker";
 import { Volume } from "gql/generated/types";
+import { HostExpirationField } from "pages/spawn/spawnHost/HostExpirationField";
 
 const { Option } = Select;
 const { TextArea } = Input;
@@ -34,39 +32,11 @@ export const HostDetailsForm: React.FC<HostDetailsFormProps> = ({
   volumes,
 }) => {
   const {
-    noExpiration,
-    expiration,
     hasUserDataScript,
     userDataScript,
     isVirtualWorkstation,
     volume,
   } = data;
-
-  const updateDate = (_, dateString) => {
-    // This functions take in the date string returned from the datePicker component
-    // It uses the date string since it comes in the format yyyy-mm-dd and it can be combined with the
-    // value from the timePicker since it only returns a subset of the date and wont overwrite the entire date
-
-    const year = dateString.slice(0, 4);
-    const month = dateString.slice(5, 7);
-    const date = dateString.slice(8, 10);
-    const updatedTime = set(expiration, { year, month, date });
-    onChange({ ...data, expiration: updatedTime });
-  };
-
-  const updateTime = (_, dateString) => {
-    // This functions take in the time string returned from the timePicker component
-    // It uses the date string since it comes in the format hh-mm-ss and it can be combined with the
-    // value from the timePicker since it only returns a subset of the date and wont overwrite the entire date
-
-    const hours = dateString.slice(0, 2);
-    const minutes = dateString.slice(3, 5);
-    const seconds = dateString.slice(7, 9);
-    const updatedTime = set(expiration, { hours, minutes, seconds });
-    onChange({ ...data, expiration: updatedTime });
-  };
-
-  const disabledDate = (current) => current < Date.now();
 
   return (
     <Container>
@@ -87,35 +57,7 @@ export const HostDetailsForm: React.FC<HostDetailsFormProps> = ({
         onChange={(e) => onChange({ ...data, userDataScript: e.target.value })}
       />
       <SectionContainer>
-        <SectionLabel>Expiration</SectionLabel>
-        <FlexColumnContainer>
-          <InputLabel htmlFor="hostDetailsDatePicker">Date</InputLabel>
-          <DatePicker
-            id="hostDetailsDatePicker"
-            onChange={updateDate}
-            disabled={noExpiration}
-            disabledDate={disabledDate}
-            value={expiration}
-          />
-        </FlexColumnContainer>
-        <PaddedBody>&amp;</PaddedBody>
-        <FlexColumnContainer>
-          <InputLabel htmlFor="hostDetailsTimePicker">Time</InputLabel>
-          <TimePicker
-            onChange={updateTime}
-            disabled={noExpiration}
-            disabledDate={disabledDate}
-            value={expiration}
-          />
-        </FlexColumnContainer>
-        <PaddedBody> or </PaddedBody>
-        <PaddedCheckbox
-          label="Never"
-          checked={noExpiration}
-          onChange={(e) =>
-            onChange({ ...data, noExpiration: e.target.checked })
-          }
-        />{" "}
+        <HostExpirationField data={data} onChange={onChange} />
       </SectionContainer>
 
       {isVirtualWorkstation && (
@@ -179,10 +121,6 @@ const Container = styled(FlexColumnContainer)`
 const PaddedBody = styled.span`
   padding-left: 15px;
   padding-right: 15px;
-  margin-top: 22px;
-`;
-
-const PaddedCheckbox = styled(Checkbox)`
   margin-top: 22px;
 `;
